api/pyproject.toml
[tool.poetry]
name = "vending"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.9.10"
fastapi = {extras = ["all"], version = "^0.70.0"}
python-multipart = "^0.0.5"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
pydantic-to-typescript = "^1.0.8"
SQLAlchemy = "^1.4.39"

[tool.poetry.dev-dependencies]
black = "^22.6.0"
flake8 = "^4.0.1"
pytest = "^7.1.2"
pytest-fastapi-deps = "^0.2.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
